fix mem leak: destroy selections and borders when overlays are destroyed
otherwise selections and borders keep hanging even after HOT was destroyed

this commit also adds some Errors that are thrown when an object is destroyed more than once. This will happen if a reference to a destroyed object is not correctly removed.

+++ src/3rdparty/walkontable/src/border.js
@@ -548,6 +548,10 @@ class Border {
    * Cleans up all the DOM state related to a Border instance. Call this prior to deleting a Border instance.
    */
   destroy() {
+    if (this.destroyed) {
+      throw new Error('This Border was already destroyed');
+    }
+    this.destroyed = true;
     this.eventManager.destroyWithOwnEventsOnly();
     this.forAllDomElements(elem => elem.parentNode.removeChild(elem));
   }
